+++ R/item_reliability.R
@@ -11,22 +11,28 @@
 #' @param digits Amount of digits for returned values.
 #' @param verbose Toggle warnings and messages.
 #'
-#' @return A data frame with the corrected item-total correlations
-#' (*item discrimination*, column `item_discrimination`) and Cronbach's Alpha
-#' (if item deleted, column `alpha_if_deleted`) for each item of the scale, or
-#' `NULL` if data frame had too less columns.
+#' @return A data frame with the item-total correlations (column
+#' `Item_Total_Correlation`), corrected item-total correlations (*item
+#' discrimination*, column `Discrimination`) and Cronbach's Alpha (if item
+#' deleted, column `Alpha_if_deleted`) for each item of the scale, or `NULL` if
+#' data frame had too less columns.
+#'
+#' @inherit check_itemscale note references
 #'
 #' @details
-#' This function calculates the item discriminations (corrected item-total
-#' correlations for each item of `x` with the remaining items) and the
-#' Cronbach's alpha for each item, if it was deleted from the scale. The
-#' absolute value of the item discrimination indices should be above 0.2. An
-#' index between 0.2 and 0.4 is considered as "fair", while an index above 0.4
-#' (or below -0.4) is "good". The range of satisfactory values is from 0.4 to
-#' 0.7. Items with low discrimination indices are often ambiguously worded and
-#' should be examined. Items with negative indices should be examined to
-#' determine why a negative value was obtained (e.g. reversed answer categories
-#' regarding positive and negative poles).
+#' This function calculates the item-total correlations, item discriminations
+#' (corrected item-total correlations for each item of `x` with the remaining
+#' items) and the Cronbach's alpha for each item, if it was deleted from the
+#' scale. The absolute value of the item discrimination indices should be above
+#' 0.2. An index between 0.2 and 0.4 is considered as "fair", while an index
+#' above 0.4 (or below -0.4) is "good". The range of satisfactory values is from
+#' 0.4 to 0.7. Items with low discrimination indices are often ambiguously
+#' worded and should be examined. Items with negative indices should be examined
+#' to determine why a negative value was obtained (e.g. reversed answer
+#' categories regarding positive and negative poles).
+#'
+#' See [`check_itemscale()`] and [`item_discrimination()`] for more details on
+#' the interpretation of the results.
 #'
 #' @examples
 #' data(mtcars)
@@ -48,7 +54,7 @@ item_reliability <- function(x, standardize = FALSE, digits = 3, verbose = TRUE)
   # remember item (column) names for return value
   # return value gets column names of initial data frame
   df.names <- colnames(x)
-  ret.df <- NULL
+  out <- NULL
 
   # check for minimum amount of columns can't be less than 3, because the
   # reliability test checks for Cronbach's alpha if a specific item is deleted.
@@ -70,23 +76,95 @@ item_reliability <- function(x, standardize = FALSE, digits = 3, verbose = TRUE)
     )
 
     # calculate corrected total-item correlation
-    totalCorr <- item_discrimination(
+    totalCorrCorrected <- item_discrimination(
       x,
       standardize = FALSE,
       verbose = verbose
     )
 
-    ret.df <- data.frame(
-      term = df.names,
-      alpha_if_deleted = round(cronbachDeleted, digits),
-      item_discrimination = round(totalCorr$Discrimination, digits),
+    # calculate total-item correlation
+    totalCorr <- item_totalcor(
+      x,
+      standardize = FALSE,
+      verbose = FALSE
+    )
+
+    out <- data.frame(
+      Item = df.names,
+      Alpha_if_deleted = round(cronbachDeleted, digits),
+      Item_Total_Correlation = round(totalCorr$Item_Total_Correlation, digits),
+      Discrimination = round(totalCorrCorrected$Discrimination, digits),
       stringsAsFactors = FALSE
     )
+
+    attr(out, "item_intercorrelation") <- item_intercor(x)
+    attr(out, "cronbachs_alpha") <- cronbachs_alpha(x)
+    class(out) <- c("item_reliability", "data.frame")
   } else {
     insight::format_warning(
       "Data frame needs at least three columns for reliability-test."
     )
   }
 
-  ret.df
+  out
+}
+
+
+# methods --------------------------------------
+
+#' @export
+print.item_reliability <- function(x, ...) {
+  out <- insight::format_table(x, ...)
+  # format column names
+  colnames(out) <- c("Item", "Alpha if deleted", "Total Correlation", "Discrimination")
+
+  # add attributes for table caption and footer
+  attr(out, "table_caption") <- c("# Item Reliability", "blue")
+  attr(out, "table_footer") <- c(sprintf(
+    "\nMean inter-item-correlation = %.3f  Cronbach's alpha = %.3f",
+    attributes(out)$item_intercorrelation,
+    attributes(out)$cronbachs_alpha
+  ), "yellow")
+
+  cat(insight::export_table(out, ...))
+  invisible(x)
+}
+
+
+#' @export
+print_md.item_reliability <- function(x, ...) {
+  .print_item_reliability(x, format = "markdown", ...)
+}
+
+
+#' @export
+print_html.item_reliability <- function(x, ...) {
+  .print_item_reliability(x, format = "html", ...)
+}
+
+
+# helper ---------------------------------------
+
+.print_item_reliability <- function(x, format = "markdown", ...) {
+  out <- insight::format_table(x, ...)
+  # format column names
+  colnames(out) <- c("Item", "Alpha if deleted", "Total Correlation", "Discrimination")
+
+  # add attributes for table caption and footer
+  caption <- "Item Reliability"
+  footer <- sprintf(
+    "Mean inter-item-correlation = %.3f  Cronbach's alpha = %.3f",
+    attributes(out)$item_intercorrelation,
+    attributes(out)$cronbachs_alpha
+  )
+
+  insight::export_table(
+    out,
+    caption = caption,
+    footer = footer,
+    format = format,
+    missing = "<NA>",
+    align = "firstleft",
+    zap_small = TRUE
+  )
 }