+++ NEWS.md
@@ -4,10 +4,22 @@
 
 * `item_omega()`, to calculate the McDonald's Omega reliability coefficient.
 
+* `item_totalcor()` calculates the total correlation of an item with the
+  sum of all other items in a scale. If `corrected = TRUE`, the total
+  correlation is corrected for the number of items in the scale (which is
+  equivalent to `item_discrimination()`).
+
+* Column names of `item_reliability()` were changed to be in line with the
+  _easystats_ naming convention and to be consistent with the output of other
+  related functions.
+
 ## Changes
 
 * `check_itemscale()` now work with factor analysis results, from
   `parameters::factor_analysis()`.
+
+* `item_reliability()` now includes the item-total correlation, and information
+  about Cronbach's alpha and mean inter-item correlation in the printed output.
 
 * `cronbachs_alpha()` now work with factor analysis results, from
   `parameters::factor_analysis()`.
@@ -20,7 +32,7 @@
 
 * `item_alpha()` was added as an alias for `cronbachs_alpha()`.
 
-* `item_totalcor()` was added as an alias for `item_discrimination()`.
+* Further functions get a `display()`, `print_md()` and `print_html()` method.
 
 ## Bug fixes